Create setupMaze.m
This function makes a matrix of certain dimensions using the point class (which will be explained later) and sets up the display of walls inside the maze. It calls 4 parameters: maze, walls, position, and end position and asks for an input of size to out put the 4 parameters. This will make the maze depending on size so it can create a certain number of walls.