Create moveMaze.m
This function deals with the movement of the icon if the maze were to be moveable (except my code only prints mazes). It calls the validate function and sees where any possible walls can be placed to fill the spaces. This makes the maze look more like a maze and is made based upon the user input of difficulty.