Create adjustEnd.m
This function seals the deal by linking the end part of the maze to rest of the maze. It looks in all directions to see if there are any spaces to build walls by calling the validate function. The function calls a maze and end position and outputs the maze display.